[CMake] Windows: Set _WIN32_WINNT (Windows SDK API compat version)
This provides two options for explicitly setting _WIN32_WINNT for Windows
builds (which sets the Windows SDK API compatibility level):

1.) WZ_WIN32_API_COMPAT_VERSION

By defining this when running CMake configure, a custom Windows SDK API
compatibility level can be specified for appropriately setting `_WIN32_WINNT`.

This setting does _not_ impact which Windows SDK is used (if multiple versions
are present). It only impacts the definition of `_WIN32_WINNT`.

The expected format is the same as that of CMAKE_SYSTEM_VERSION. Examples:
- Windows 7:   "6.1"
- Windows 8:   "6.2"
- Windows 8.1: "6.3"

2.) CMAKE_SYSTEM_VERSION

If WZ_WIN32_API_COMPAT_VERSION is _not_ defined (or is empty), and
CMAKE_SYSTEM_VERSION is explicitly set, it will be used to set an appropriate
value for _WIN32_WINNT.

Note: Setting CMAKE_SYSTEM_VERSION can impact other build configuration, like
which version of the Windows SDK is used (if multiple versions are present).
See the CMake documentation.