TextDraw: Fix text shaping width calculations (WIP)
We need to differentiate between the width needed for the rasterized buffer, and the *layout* width of the text (i.e. the accumulation of x_advance, or the next position on the line at which a character could start)

WzText::width()/height() should return the _layout_ width/height, as should getTextMetrics, iv_GetTextWidth/Height, etc. What is currently treated as the "width" should simply be the dimensions of the rasterized text image (internal-only - used for drawing only). The "offset" is also only used for drawing. Publicly-exposed functions should return layout dimensions (which may be bigger in the case of a string with spaces on the ends).